+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Changed
+- Bearer token is propagated through the `httpx_client`.
 - Use `.model_dump(exclude_defaults=True)` for auto-generated classes dumping.
 - Get morpho tool and resolve entities now uses entitycore.
 - Remove usage of user journey in suggestion after first message.